requirements.txt
scikit-learn==1.3.0
zipp==3.16.2
pandas==2.0.3
requests==2.31.0
xgboost==2.0.0
optuna==3.3.0
mlflow==2.7.0
apache-airflow==2.7.1
apache-airflow-providers-apache-spark==4.9.0
pyspark==3.3.0